#8 second relocation of code
